move function to Move class
